+++ src/Presentation/SmartStore.Web/Controllers/ShoppingCartController.cs
@@ -244,8 +244,8 @@ namespace SmartStore.Web.Controllers
 
             //codehint:sm-add
             model.MediaDimensions = _mediaSettings.CartThumbPictureSize;
-            //TODO: add setting
             model.DisplayDeliveryTime = _shoppingCartSettings.ShowDeliveryTimes;
+            model.DisplayShortDesc = _shoppingCartSettings.ShowShortDesc;
             model.IsEditable = isEditable;
             model.ShowProductImages = _shoppingCartSettings.ShowProductImagesOnShoppingCart;
             model.ShowSku = _catalogSettings.ShowProductSku;
@@ -435,6 +435,7 @@ namespace SmartStore.Web.Controllers
             foreach (var sci in cart)
             {
                 sci.Product.MergeWithCombination(sci.AttributesXml);
+                
                 var cartItemModel = new ShoppingCartModel.ShoppingCartItemModel()
                 {
                     Id = sci.Id,
@@ -445,8 +446,15 @@ namespace SmartStore.Web.Controllers
                     Quantity = sci.Quantity,
                     AttributeInfo = _productAttributeFormatter.FormatAttributes(sci.Product, sci.AttributesXml),
                     IsShipEnabled = sci.Product.IsShipEnabled,
-                    DeliveryTime = _deliveryTimeService.GetDeliveryTimeById(sci.Product.DeliveryTimeId.GetValueOrDefault())
+                    ShortDesc = sci.Product.ShortDescription
                 };
+
+                var deliveryTime = _deliveryTimeService.GetDeliveryTimeById(sci.Product.DeliveryTimeId.GetValueOrDefault());
+                if (deliveryTime != null) 
+                { 
+                    cartItemModel.DeliveryTimeName = deliveryTime.GetLocalized(x => x.Name);
+                    cartItemModel.DeliveryTimeHexValue = deliveryTime.ColorHexValue;
+                }
 
                 //allowed quantities
                 var allowedQuantities = sci.Product.ParseAllowedQuatities();